Moved server class command from data_structures
